added verification page and fixed profile picture
profile picture can be changed now and got rid of the default image as well as fixed the profile picture not changing in the header tab. There is still a quick flash when entering the profile page but you cant really notice it as much as before. (default image was red so the flash was more visable)